+++ lib/src/main/java/com/formance/formance_sdk/models/shared/Monetary.java
@@ -6,14 +6,16 @@ package com.formance.formance_sdk.models.shared;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigInteger;
+
 public class Monetary {
     /**
      * The amount of the monetary value.
      */
     @JsonProperty("amount")
-    public Long amount;
+    public BigInteger amount;
 
-    public Monetary withAmount(Long amount) {
+    public Monetary withAmount(BigInteger amount) {
         this.amount = amount;
         return this;
     }
@@ -29,7 +31,7 @@ public class Monetary {
         return this;
     }
     
-    public Monetary(@JsonProperty("amount") Long amount, @JsonProperty("asset") String asset) {
+    public Monetary(@JsonProperty("amount") BigInteger amount, @JsonProperty("asset") String asset) {
         this.amount = amount;
         this.asset = asset;
   }
